perf: split modal dialogs code away from main bundle
