+++ packages_rs/nextclade-web/package.json
@@ -175,6 +175,7 @@
     "@testing-library/user-event": "14.1.1",
     "@types/classnames": "2.3.0",
     "@types/compare-versions": "3.3.0",
+    "@types/compression": "1.7.2",
     "@types/compression-webpack-plugin": "9.0.0",
     "@types/connect-history-api-fallback": "1.3.5",
     "@types/copy-webpack-plugin": "6.4.3",
@@ -225,6 +226,7 @@
     "allow-methods": "3.1.0",
     "babel-plugin-parameter-decorator": "1.0.16",
     "babel-plugin-transform-typescript-metadata": "0.3.2",
+    "compression": "1.7.4",
     "compression-webpack-plugin": "9.2.0",
     "connect-history-api-fallback": "1.6.0",
     "conventional-changelog-cli": "2.2.2",
